skraafoto-api/requirements.txt
requests
pillow
python-dotenv
folium
pyproj
matplotlib
opencv-python-headless
xmltodict
